fix: fixed auth check
