+++ innws-stack-overflow/src/features/auth/api/authAPI.ts
@@ -1,0 +1,16 @@
+import { api } from "../../../shared/api/api";
+
+export const authUser = async() => {
+  try {
+    const response = await api.get("/auth");
+
+    if (response.status === 200) {
+      return true;
+    } else {
+      return false;
+    }
+  } catch (error) {
+    console.error("Auth check failed", error);
+    return false;
+  }
+}
