+++ innws-stack-overflow/src/shared/hooks/useAuth.tsx
@@ -1,0 +1,12 @@
+import { createContext, useContext } from "react";
+import { AuthContextType } from "../types/auth";
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error("useAuth must be used within an AuthProvider");
+  }
+  return context;
+};
