--- innws-stack-overflow/src/shared/utils/authUtils.tsx
@@ -1,22 +1,0 @@
-import { createContext, useContext, useState, ReactNode } from "react";
-import { AuthContextType } from "../types/auth";
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const AuthProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
-  const [isAuth, setIsAuth] = useState<boolean>(!!localStorage.getItem("user"));
-
-  return (
-    <AuthContext.Provider value={{ isAuth, setIsAuth }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-};
